* remotes/origin/master:
  Added firebird-driver.conf for usage in my own env.
  Added/Updated bugs\core_5793_test.py. adjusted for usage together with QA_GLOBALS dict (read common settings from act.files_dir/test_config.ini)
  Added/Updated bugs\core_5796_test.py. adjusted for usage together with QA_GLOBALS dict (read common settings from act.files_dir/test_config.ini)
  Added/Updated bugs\core_6048_test.py. adjusted for usage together with QA_GLOBALS dict (read common settings from act.files_dir/test_config.ini)
  Added/Updated bugs\core_5808_test.py. adjusted for usage together with QA_GLOBALS dict (read common settings from act.files_dir/test_config.ini)
  Added/Updated bugs\core_5831_test.py. adjusted for usage together with QA_GLOBALS dict (read common settings from act.files_dir/test_config.ini)
  Added/Updated bugs\core_6071_test.py. adjusted for usage together with QA_GLOBALS dict (read common settings from act.files_dir/test_config.ini)
  Added/Updated bugs\core_6163_test.py. adjusted for usage together with QA_GLOBALS dict (read common settings from act.files_dir/test_config.ini)
  Added ENCRYPTION_BADKEY setting.
  Raise exception instead key error when sample_dir cannot be determined
  Added/Updated bugs\core_6947_test.py. adjusted for usage together with QA_GLOBALS dict (read common settings from act.files_dir/test_config.ini)
  Fix using explicit credentials in Action.trace()

+++ firebird/qa/plugin.py
@@ -377,6 +377,7 @@ def pytest_configure(config):
             result = 'SuperClassic' if f1 == f2 else 'SuperServer'
     _vars_['server-arch'] = result
     # get sampleDB directory
+    _vars_['sample_dir'] = None
     with connect('employee') as con, \
          connect_server(_vars_['server'], user='SYSDBA', password=_vars_['password']) as srv:
         for db in srv.info.attached_databases:
@@ -384,6 +385,8 @@ def pytest_configure(config):
             if db_path.name.lower() == 'employee.fdb':
                 _vars_['sample_dir'] = db_path.parent
     # Create QA subdirectory in samle db directory
+    if _vars_['sample_dir'] is None:
+        raise Exception("Cannot determine the 'sample_dir'")
     samle_qa = _vars_['sample_dir'] / 'qa'
     remove_dir(samle_qa)
     samle_qa.mkdir(exist_ok=True)
@@ -1094,7 +1097,8 @@ def trace_thread(act: Action, b: Barrier, cfg: List[str], output: List[str], kee
         encoding: Encoding for trace session output.
         encoding_errors: Error handler for trace session output encoding.
     """
-    with act.connect_server(encoding=encoding, encoding_errors=encoding_errors) as srv:
+    with act.connect_server(encoding=encoding, encoding_errors=encoding_errors,
+                            user=user, password=password) as srv:
         output.append(srv.trace.start(config='\n'.join(cfg)))
         b.wait()
         for line in srv:
